+++ docs/changelog.rst
@@ -35,6 +35,7 @@ New Features
 - QoL method :meth:`~gatoh.model.ABModel.add_relationships_to_hierarchy`
 - Defined multiple global constants throughout modules to clarify "magic numbers"
 - Prototype for ability to add "link functions" to model parameters
+- Prototype for the ability to track specific model parameters throughout iterations within the :class:`~gatoh.logging.GATOHLogger`
 
 Fixes
 -----